refactor(tts): name the plain-local guard and fold the OmniVoice tests
Two cleanups left over from #141.

omnivoice is a member of LOCAL_TTS_PROVIDERS, so isLocal alone sweeps it
in. URL normalization wants that, but the connection hints and HTTP error
text do not, which left an isLocal && !isOmniVoice condition that every
future local branch would have to remember. Names it isPlainLocal once
and uses it at both sites.

omnivoice.test.ts had no matching module after the fold and carried a
duplicate copy of the AudioContext scaffolding. Its cases move into
openai-tts.test.ts, plus a new case asserting OmniVoice and local-tts
keep distinct hints and error text.

Behaviour unchanged: re-ran the mutation-validated differential against
the pre-fold classes, 12764 comparisons, 0 divergences.
