+++ src/lib/services/tts/openai-tts.test.ts
@@ -174,3 +174,117 @@ test('sends Authorization only when an API key is present', async () => {
 	assert.equal(seen[0]?.Authorization, 'Bearer secret');
 	assert.equal('Authorization' in (seen[1] ?? {}), false);
 });
+
+// --- OmniVoice, served by this same client since the fold in #141 ---
+
+test('fetchAudioBuffer sends WAV request with base fields and no API key', async () => {
+	const requests: { url: string; body: Record<string, unknown>; headers: Record<string, string> }[] = [];
+	// @ts-expect-error global fetch mock
+	globalThis.fetch = (url: string, init: RequestInit) => {
+		requests.push({ url, body: parseBody(init), headers: init.headers as Record<string, string> });
+		return mockFetchResponse();
+	};
+
+	const tts = new OpenAITTS({
+		provider: 'omnivoice',
+		voiceId: 'alloy',
+		language: 'de',
+		speed: 1.2
+	});
+
+	await tts.fetchAudioBuffer('Hallo Welt.');
+
+	assert.equal(requests.length, 1);
+	assert.equal(requests[0].url, 'http://localhost:8881/v1/audio/speech');
+	assert.equal(requests[0].body.model, 'omnivoice');
+	assert.equal(requests[0].body.input, 'Hallo Welt.');
+	assert.equal(requests[0].body.voice, 'alloy');
+	assert.equal(requests[0].body.language, 'de');
+	assert.equal(requests[0].body.speed, 1.2);
+	assert.equal(requests[0].body.response_format, 'wav');
+	assert.equal('Authorization' in requests[0].headers, false);
+});
+
+test('fetchAudioBuffer overrides language and speed from stream options', async () => {
+	const requests: { body: Record<string, unknown> }[] = [];
+	// @ts-expect-error global fetch mock
+	globalThis.fetch = (_url: string, init: RequestInit) => {
+		requests.push({ body: parseBody(init) });
+		return mockFetchResponse();
+	};
+
+	const tts = new OpenAITTS({
+		provider: 'omnivoice',
+		voiceId: 'alloy',
+		language: 'en',
+		speed: 1.0
+	});
+
+	await tts.fetchAudioBuffer('Hello.', { language: 'es', speed: 0.9 });
+
+	assert.equal(requests[0].body.language, 'es');
+	assert.equal(requests[0].body.speed, 0.9);
+});
+
+test('speak returns a source and analyser', async () => {
+	globalThis.fetch = () => mockFetchResponse();
+
+	const tts = new OpenAITTS({
+		provider: 'omnivoice'
+	});
+
+	const result = await tts.speak('Hello world.');
+	assert.ok(result.source);
+	assert.ok(result.analyser);
+});
+
+test('connection error surfaces a local OmniVoice hint', async () => {
+	globalThis.fetch = () => Promise.reject(new Error('Failed to fetch'));
+
+	const tts = new OpenAITTS({
+		provider: 'omnivoice',
+		baseUrl: 'http://localhost:8880/v1'
+	});
+
+	await assert.rejects(tts.fetchAudioBuffer('Hello.'), /OmniVoice proxy/);
+});
+
+test('HTTP error surfaces providerErrorMessage-style detail', async () => {
+	globalThis.fetch = () =>
+		Promise.resolve({
+			ok: false,
+			status: 400,
+			json: () => Promise.resolve({ error: { message: 'bad request' } })
+		} as Response);
+
+	const tts = new OpenAITTS({
+		provider: 'omnivoice'
+	});
+
+	await assert.rejects(tts.fetchAudioBuffer('Hello.'), /OmniVoice error 400: bad request/);
+});
+
+test('OmniVoice keeps its own hint when a plain local server would not', async () => {
+	// omnivoice is in LOCAL_TTS_PROVIDERS, so a bare isLocal check would wrongly
+	// hand it the local-tts messages. Guard the two apart explicitly.
+	globalThis.fetch = () => Promise.reject(new TypeError('Failed to fetch'));
+	await assert.rejects(
+		() => new OpenAITTS({ provider: 'omnivoice' }).fetchAudioBuffer('x'),
+		/OmniVoice proxy/
+	);
+	await assert.rejects(
+		() => new OpenAITTS({ provider: 'local-tts' }).fetchAudioBuffer('x'),
+		/local TTS server/i
+	);
+
+	globalThis.fetch = () =>
+		Promise.resolve({ ok: false, status: 400, json: () => Promise.resolve({}) } as Response);
+	await assert.rejects(
+		() => new OpenAITTS({ provider: 'omnivoice' }).fetchAudioBuffer('x'),
+		/OmniVoice error 400/
+	);
+	await assert.rejects(
+		() => new OpenAITTS({ provider: 'local-tts' }).fetchAudioBuffer('x'),
+		/Local TTS server returned 400/
+	);
+});
